refactor all assistant chat demos
